create separate entry points for node and browser
