+++ anycode/server/src/browser/main.ts
@@ -1,0 +1,29 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { BrowserMessageReader, BrowserMessageWriter, createConnection } from 'vscode-languageserver/browser';
+import { SymbolInfoStorage } from '../features/symbolIndex';
+import { IStorageFactory, startServer } from '../server';
+import { IndexedDBSymbolStorage } from './storage';
+
+const messageReader = new BrowserMessageReader(self);
+const messageWriter = new BrowserMessageWriter(self);
+
+const connection = createConnection(messageReader, messageWriter);
+
+const factory: IStorageFactory = {
+	async create(name) {
+		const result = new IndexedDBSymbolStorage(name);
+		await result.open();
+		return result;
+	},
+	async destroy(obj: SymbolInfoStorage) {
+		if (obj instanceof IndexedDBSymbolStorage) {
+			await obj.close();
+		}
+	}
+};
+
+startServer(connection, factory);
